# static/swagger/swagger.yaml
...
swagger: "2.0"
info:
  version: "1.0.0"
  title: Mechanic Workshop API
  description: API documentation for Customers, Mechanics, Service Tickets, and Inventory.
# API server configuration specifying host, base path, and HTTP scheme
host: localhost:5000
basePath: /api
schemes:
  - http
# Security definition for JWT Bearer token authentication via Authorization header
securityDefinitions:
  BearerAuth:
    type: apiKey
    name: Authorization
    in: header
    description: >
      JWT Token authentication. Use 'Bearer <your-token>' format.
# Tags categorizing API endpoints by resource type
...
paths:
...
  /service-tickets:
    post:
      tags:
        - Service Tickets
      summary: Create Service Ticket
      description: "Creates a new service ticket."
      security:
        - BearerAuth: []
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/ServiceTicketInput'
            example:
              description: "Oil change and brake inspection"
              customer_id: 1
              vin: "1HGCM82633A123456"
      responses:
        201:
          description: "Service ticket created successfully"
          schema:
            type: object
            properties:
              message:
                type: string
              ticket:
                type: object
                properties:
                  id:
                    type: integer
                  customer_id:
                    type: integer
                  description:
                    type: string
                  vin:
                    type: string
                  mechanics:
                    type: array
                    items:
                      type: object
                  parts:
                    type: array
                    items:
                      type: object
          examples:
            application/json:
              message: "Service ticket created successfully."
              ticket:
                id: 101
                customer_id: 1
                description: "Oil change and brake inspection"
                vin: "1HGCM82633A123456"
                mechanics: []
                parts: []
# Retrieve a paginated list of all service tickets (requires authentication)
  /service-tickets/:
    get:
      tags:
        - Service Tickets
      summary: Find All Service Tickets (Paginated)
      description: Returns a paginated list of service tickets.
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          type: integer
          required: false
          description: "Page number (default: 1)"
        - name: per_page
          in: query
          type: integer
          required: false
          description: "Results per page (default: 10)"
      responses:
        200:
          description: A paginated list of service tickets.
          schema:
            type: object
            properties:
              service_tickets:
                type: array
                items:
                  $ref: '#/definitions/ServiceTicket'
              total:
                type: integer
              pages:
                type: integer
              current_page:
                type: integer
# Retrieve service ticket details by ID (requires authentication)
  /service-tickets/{id}:
    get: 
      tags: [Service Tickets]
      summary: Find Service Ticket by ID
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          type: integer
      responses:
        200:
          schema:
            $ref: '#/definitions/ServiceTicket'
# Update a service ticket's details by ID (requires authentication)
    put:
      tags:
        - Service Tickets
      summary: Update service ticket by ID
      description: "Updates an existing service ticket."
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          type: integer
          required: true
          description: ID of the service ticket to update
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/ServiceTicketInput'
            example:
              description: "Brake pad replacement"
              customer_id: 1
              vin: "1HGCM82633A123456"
      responses:
        200:
          description: "Service ticket updated successfully"
          schema:
            $ref: '#/definitions/ServiceTicket'
          examples:
            application/json:
              id: 101
              customer_id: 1
              description: "Brake pad replacement"
              vin: "1HGCM82633A123456"
# Delete a service ticket by ID (requires authentication)
    delete:
      tags: [Service Tickets]
      summary: Delete Service Ticket by ID
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          type: integer
      responses:
        204:
          description: Service Ticket deleted successfully
# Assign a mechanic to a service ticket by their IDs (requires authentication)
  /service-tickets/{ticket_id}/assign-mechanic/{mechanic_id}:
    put:
      tags: [Service Tickets]
      summary: Assign Mechanic to Service Ticket by ID
      security:
        - BearerAuth: []
      parameters:
        - name: ticket_id
          in: path
          required: true
          type: integer
        - name: mechanic_id
          in: path
          required: true
          type: integer
      responses:
        200:
          description: Mechanic assigned successfully
# Remove a mechanic from a service ticket by their IDs (requires authentication)
  /service-tickets/{ticket_id}/remove-mechanic/{mechanic_id}:
    delete:
      tags: [Service Tickets]
      summary: Remove Mechanic from Service Ticket by ID
      security:
        - BearerAuth: []
      parameters:
        - name: ticket_id
          in: path
          required: true
          type: integer
        - name: mechanic_id
          in: path
          required: true
          type: integer
      responses:
        204:
          description: Mechanic removed successfully
# Add or remove mechanics on a service ticket by ID (requires authentication)
  /service-tickets/{ticket_id}/edit:
    put:
      tags: [Service Tickets]
      summary: Edit Mechanics on Service Ticket by ID
      description: Add or remove mechanics from a service ticket
      security:
        - BearerAuth: []
      consumes:
        - application/json
      parameters:
        - name: ticket_id
          in: path
          required: true
          type: integer
        - in: body
          name: body
          required: true
          schema:
            type: object
            properties:
              add_ids:
                type: array
                items:
                  type: integer
                example: [0]
              remove_ids:
                type: array
                items:
                  type: integer
                example: [0]
            required:
              - add_ids
              - remove_ids
      responses:
        200:
          description: Service ticket edited successfully
          schema:
            $ref: '#/definitions/ServiceTicket'
# Add parts to a service ticket by ID (requires authentication)
  /service-tickets/{ticket_id}/add-parts:
    post:
      tags: [Service Tickets]
      summary: Add Parts to Service Ticket by ID
      security:
        - BearerAuth: []
      consumes:
        - application/json
      parameters:
        - name: ticket_id
          in: path
          required: true
          type: integer
        - in: body
          name: body
          required: true
          schema:
            type: object
            properties:
              part_ids:
                type: array
                items:
                  type: integer
                example: [0]
      responses:
        200:
          description: Parts added successfully
# Retrieve all parts associated with a service ticket by ID (requires authentication)
  /service-tickets/{ticket_id}/parts:
    get:
      tags: [Service Tickets]
      summary: Find Parts on Service Ticket by ID
      security:
        - BearerAuth: []
      parameters:
        - name: ticket_id
          in: path
          required: true
          type: integer
      responses:
        200:
          schema:
            type: array
            items:
              $ref: '#/definitions/Inventory'
# Remove a part from a service ticket by ID (requires authentication)
  /service-tickets/{ticket_id}/remove-part/{part_id}:
    delete:
      tags: [Service Tickets]
      summary: Remove Parts from Service Ticket by ID
      security:
        - BearerAuth: []
      parameters:
        - name: ticket_id
          in: path
          required: true
          type: integer
        - name: part_id
          in: path
          required: true
          type: integer
      responses:
        204:
          description: Part removed successfully
# Create a new inventory part (requires authentication)
...
definitions:
...
  ServiceTicket:
    type: object
    properties:
      id:
        type: integer
      description:
        type: string
      customer_id:
        type: integer
      vin:
        type: string
    required:
      - id
      - description
      - customer_id
      - vin   
# Input model for creating or updating a service ticket with required fields
  ServiceTicketInput:
    type: object
    required: [description, customer_id, vin]
    properties:
      description:
        type: string
      customer_id:
        type: integer
      vin:
        type: string
        description: Vehicle Identification Number 
        example: "1HGCM82633A123456"
# Inventory object model representing a part’s details in stock
  Inventory:
    type: object
    properties:
      id:
        type: integer
      name:
        type: string
      sku:
        type: string
      description:
        type: string
      price:
        type: number
    required:
      - name
      - sku
      - description
      - price
# Input model for creating or updating an inventory part with required fields
...